* Fix the problem that the default screensaver * cannot automatically switch the background on v10.
