+++ Common/commonfunc.h
@@ -1,0 +1,17 @@
+#ifndef COMMONFUNC_H
+#define COMMONFUNC_H
+
+#include <QString>
+
+
+bool ispicture(QString filepath);
+QString getSystemVersion();
+QString getSystemDistrib();
+
+class commonFunc
+{
+public:
+    commonFunc();
+};
+
+#endif // COMMONFUNC_H
